Integration test to verify all tests are run on the same thread under @RunWith.
Summary:
Updated `DelegateRunnerWithTimeout` so that it maintains the default behavior of
JUnit where all tests are run on the same thread. We observed some test cases in
the wild that relied on this behavior. Specifically, Robolectric tests rely on this
due to a check in `org.robolectric.shadows.ShadowLooper`.

Test Plan:
Add a test to the `RunWithDefaultTimeoutIntegrationTest` test case to exercise the fix.
Made `executor` an instance rather than a `DelegateRunnerWithTimeout` field and verified
that `RunWithDefaultTimeoutIntegrationTest#testAllTestsForRunWithAreRunOnTheSameThread()`
failed.