python_library should not flatten libraries.
Summary:
Multi-file python libraries are often contained in a directory and
imported by that directory's name. Flattening the hierarchy will render that
module unimportable (or worse, possibly import only part of it). By copying the
unflattened subtree, we preserve the semantics of Python's module system while
retaining the isolation property.