Refactor CommandLineBuildTargetNormalizer to use pass-from-above.
Summary:
`CommandLineBuildTargetNormalizer` is constructed lazily by and made available through
`AbstractCommandOptions`.

This diff also deletes `BuildTargetNormalizer`, which apparently was unused.

This reduces the number of `Function` objects that need to be created.
This will make it simpler to add a normalize() method that takes a
single string, which will be done in a follow-up diff that uses it.

Test Plan: Sandcastle builds.