Update the strings asset file format for easier parsing.
Summary:
The new file format makes it so that parsing sections of the file one at a time
becomes easier.
Also, switch some of the `byte` and `short` to `int` to make it more flexible.
Intentionally left the `byte` data type in plurals since the number of plural
categories is guaranteed to be <= 6.

To make it easier to maintain the unit test, I changed the file format to
something easy to understand/edit.

Test Plan: ant test