Refactor graph enhancement logic into AndroidBinaryGraphEnhancer.
Summary:
This makes the graph enhancement logic easier to unit test, and helps
ensure that it does not modify the `AndroidBinaryRule.Builder` from which
it is used. Its only side-effect is that it updates the `BuildRuleResolver`
that is passed in, but now that distinction is clearer.

Test Plan: Sandcastle builds.