Made some improvements to the set of fakes we use for testing.
Summary:
Currently, we frequently mock BuildContext in unit tests for lack of a
good fake. This introduces a utility to set many of BuildContext.Builder's
fields to a reasonable default for tests.

It also adds a listener to the BuckEventBus used for tests that will
print errors to the console.

I also refactored some things to make it easier to specify a FakeFileHashCache
that can be used with a FakeRuleKeyBuilderFactory, which can in turn be used
with a FakeAbstractBuildRuleBuilderParams.

Test Plan: Sandcastle builds.