Extract an InitializeFromDisk interface.
Summary:
Separate the concern of being a BuildRule or a Buildable from the concern of
being initialized from the disk. Ultimately, this is a responsibility of the
"build engine" rather than the rule itself, so this will help make the
separation of BuildEngine from BuildRule less painful.

Test Plan:
buck test --all
Clean build of the stable branch, checking that we hit the caches.