Fix how ConstructorArgMarshaller populates an ImmutableSortedSet<SourcePath>.
Summary:
The JSON generated from `buck.py` produces file paths relative to the build file,
but Buck logic expects paths relative to the project root.
Therefore, if the type of a field in an `Arg` is `SourcePath` or `CollectionSourcePath`,
then it must resolve the strings returned by the JSON against the directory that contains
the build file in which the rule was defined.

Test Plan: Sandcastle builds.