Fix regex when downscaling from drawable/.
Summary:
Unlikely case, but if we had `drawable/x.png` and wanted to
target LDPI, the regex would have failed to match it.

Test Plan:
unit tests, still work. Mucked around with `FilterResourcesStepTest` for a bit
to see what would happen if I targeted LDPI and had `drawable` as a source.