Improve the error message for source_under_test entries that aren't dependencies.
Summary:
We're incorrectly reporting these as being the wrong type, when
in fact they're coming back as null because they're not visible to the
ruleResolver. Throw the correct exception.

Test Plan:
Delete one of these sources under test from it's associated
deps list, and confirm that this error shows correctly.