Insert canary classes in secondary .dex files when merging pre-dexed files.
Summary:
As explained in CanaryFactory, this ensures that every secondary dex file has
a "canary" class when secondary dexes are created via merging pre-dexed files.
The canary class is known to be safe to load on any system, which is how we
verify that the secondary dex has been loaded successfully.

Test Plan: Sandcastle builds.