Upload build results asynchronously from buckd.
Summary:
Call NGContext.exit() as soon as possible after the buck exit
code has been determined and upload results in the background
after the client has exited. Ensuring that the SuperConsole renders
its final frame is fiddly in this case: the Console must be created
before the BuildEventBus in the try-with-resoures block so that the
BuildEventBus is closed first and so delivers the final event to the
console before the console is closed. The console close method must
also synchronously call render to output the final frame to the
NGContext error stream before the exitcode is sent to the nailgun client.