Introduce an ignored failing test demonstrating a problem with SourcePath deps.
Summary:
It should be possible for a DescribedRule to depend on another target simply
by including a SourcePath as one of the parameters of the constructor arg.
This test case demonstrates that this isn't currently the case.

Test Plan: None. There are no logical changes, and the only test case is ignored.