Disable SuperConsole when `-v 2` or above is specified.
Summary:
When building with `-v 2` or above, the verbose output disables the
SuperConsole, which then fails to print messages posted to the event
bus.  As a workaround until this is resolved, this diff disabled the
SuperConsole if `-v 2` or above is specified.  The fallback
SimpleConsole prints the messages properly.

Test Plan: Did a build with -v 2 and saw proper error messages.