Rename `primary_dex_substrings` to `primary_dex_patterns`.
Summary:
I'm going to add support for prefix, suffix, and exact match to allow us
to have tighter patterns.  Preemptively renaming these to make that diff
easier to read.

This requires a corresponding rename diff in fbandroid when we bump
buckversion.

This feature is not used outside of Facebook, so there is no need for an
external migration plan or documentation update.

Test Plan: Built FB4A.