Include the command used to create the trace in `/traces`.
Summary:
This makes the `/traces` page easier to use, as you can see
what command was executed to produce the trace.

I will also include this information on the individual `/trace`
page in a follow-up diff.

Test Plan:
Updated `TracesHandlerTest`.  Visited the `/traces` servlet from buckd's
web server and verified that the correct command appears for the trace.