Add a low-level debug log and use it to diagnose silent test failures.
Summary:
We're seeing cases on our build servers where `buck test` is exiting
with a failure code but not printing any output that would indicate what
the problem is.  This diff adds a very low-level file log for collecting
debugging information about this issue.

I'll file a task to remove all of the calls added in this diff.  We can
either keep `DebugLog` around as a quick way to debug issues like this or
replace it with a more robust logging facility.

Test Plan:
Built locally with successful tests, unsuccessful tests, and a build
failure.  The latter two added useful messages to the log.