Improvements to logic around `BuckEventBus` shutdown.
Summary:
We have been seeing cases where test runs will report failure without
printing any failure messages to stderr.  I suspect that this is due to
the BuckEventBus not being fully shut down.  This is an attempt to
improve the diagnostic if that is the case.

`BuckEventBus.close` no longer waits for termination.  This will still be
invoked by the try-with-resources block in the event of an exception to
shut down the thread pool.  @jimp, let me know if you think we need to
wait for termination in this case.

After posting the finished event, we explicitly wait for termination,
and print an error message to the console if it times out.  If this is
what is causing the issues we've been seeing, then we'll at least be
able to confirm it.

I fixed the handling of `InterruptedException` in shutdown as well, which
should not matter here, but it's good style, and the old comment didn't
make sense.

Test Plan:
Unit tests for the successful and unsuccessful shutdown cases.  These
are really more like integration tests, but a fully-deterministic test
would not accurately test that this change is working as intended.

I also did a test run with a sleep in the SimpleConsole and saw the
error message.  Previously, it just finished without printing its full
output.