Fix order-sensitivity in `java_library_rule`.
Summary:
A subtle bug in `DefaultJavaLibraryRule`: it made us use the hash-order
list of sources instead of the sorted sources when computing the
`RuleKey`.  This caused unstable `RuleKeys` for libraries with enough
source files to cause hash collisions.

Resources had the same problem, but we didn't notice it because we don't
use resources in our Java libraries.

Test Plan: New unit test (which fails without the diff).