Fix test selectors for JUnit 4.8.
Summary: When using test selectors in fb4a, which has an older JUnit,
JUnitRunner's detection of "a test class that ran but had all its
results filtered out" needed fixing.  The behaviour appears to be
reasonable for the older version of JUnit -- testless classes do not
cause exceptions to be reported as a failing result.

Test Plan: Included integration test.