Prevent stale versions of buck from running when ant is not present.
Summary:
Previously, if the user didn't have ant installed, when buck was
auto-upgraded due to a `.buckversion` update, the the build would fail,
but leave the successful-build file in place because `ant clean` could
not be run.  Now, we manually remove that file to prevent the next buck
command from continuing with the old code.

Test Plan:
- Created a few revisions where `command ant` was replaced with
  `command ants` to simulate ant not being installed on my system.
- Checked out one of the revisions and manually built buck.
- Set buckversion to the other and tried to build fb4a.
- Prior to this change, it would fail, but running `buck build fb4a` a
  second time would succed.
- With this change, it always failed.