Simplify type traversal on `TypeCoercer`.
Summary:
Type traversals visits all the types in a type coercer.

This is used in the process of building a list of dependencies to load
before formally parsing the field itself (where it will require the
loaded dependencies). This is used to skip fields which can never
contain any dependencies.

I've replaced it with a query function for whether the type coercer has
some particular types.

Test Plan: Add unit tests.