`Builder` interface for `TestSelectorList`.
Summary:
Turn the static `TestSelectorList` builder method into its own builder
class, to allow for other things to be added to a `TestSelectorList`.

(Note: While it would be handy to use immutable collections here, we're
in a guava-free zone due to sharing this code with the `JUnitRunner`.)

Test Plan: I've converted the existing tests to use the builder.