Move dex merging into a separate buildable.
Summary:
This is similar in spirit to the change that created `UberRDotJava`.
`PreDexMerge` is the new buildable that is created during graph
enhancement and responsible for merging the pre-dexed libraries.

The immediate perf gain from this is very minor (aapt package can run in
parallel with dex merging), but it paves the way for building apks
without secondary dex files included.

Most of the code is just pulling the step creation functions out of
`AndroidBinaryRule`, but there are a few other changes like recording
outputs and saving metadata for proper caching.

Test Plan:
Minor changes to unit test.
Built and ran FB4A and Messenger (split dex).
Built and ran Instagram (single dex, with merge).
Built and ran FB4A inhouse (proguard, no merge).