Fix how we compute rule keys for build rules with exported deps.
Summary:
Repro steps:
A buck project with the following source files:

A.java:
```
public class A extends B {}
```

B.java:
```
public class B {
  private C c;
}
```

C.java:
```
public class C {}
```

BUCK:
```
java_library(
  name='a',
  srcs = [ 'A.java' ],
  exported_deps = [
  ':b',
  ],
)

java_library(
  name='b',
  srcs = [ 'B.java' ],
  deps = [
  ':c',
  ],
)

java_library(
  name='c',
  srcs = [ 'C.java' ],
)
```

Now run:
```
$ buck build //:a
// Now edit C.java (add a newline at the end so that its ABI doesn't change).
$ buck build //:a
```

Load build trace and notice C gets rebuilt and B is skipped (ABI hit),
however A gets rebuilt, which is incorrect since it should also get an
ABI hit.

The reason A does not get an ABI hit is because of the way we compute
rule keys for java library rules. A rule has an ABI match only if its
rule-key-without-deps + abi key of all its deps stays the same.
Currently, since the rule-key-without-deps of A includes the rule key of
B (an exported dep, which changes), A doesn't get an ABI hit. This
change makes it so that the rule key of B (i.e. an exported dep) is not
a part of the rule-key-without-deps (moving it to the complete rule
key).

Moving `ExportDependencies.java` to the `build_rule` BUCK rule seems
like a hack, but I am not sure how bad that is.

Also, this was the best bug report ever - thanks to @dreiss!

Test Plan: Follow the repro steps and verify that only C gets re-built.