Xcode project serializer
Summary:
This adds an in-memory representation of an Xcode project for building
an xcode project from parts, and serializing the in-memory
representation to an XML plist.

To preempt requests to change the class names so they are properly
camelcased: The class and member names mirror that of the xcode file
format itself, changing them gains conformity with the rest of the
codebase but loses conformity with the actual file format it's modeling.

Test Plan:
A simple unit test. Will add more once we have the third-party plist
library.