Introduce an "exopackage" argument for android_binary
Summary:
An "exopackage" is an APK where the important bits of it are outside of
the APK, rather than within it.  Eventually, this could include native
libraries and possibly resources and images, but right now it is just
secondary dex files.

If an android_binary has the "exopackage = True" argument, its secondary
dex files (including metadata.txt) will not be included in the APK.

Upcoming diffs will use this as an optimization for build and install
time.

Test Plan:
Sorry, tests coming in a separate diff.  I ended up creating all of my
testdata in a diff that depends on this one, and separating it out
doesn't seem worth it.