Do not fail during shutdown processing.
Summary:
Currently if a timeout occurs during shutdown a non-zero value is
returned by Buck, which can cause integration tests to erroneously fail.
This change restores the original behaviour which was to return 0 if an
exception is thrown during shutdown processing while also continuing to
report shutdown exceptions on the console to aid debugging.

Test Plan:
0) Disconnect from the network.
1) `buck test --all`