`ShellStep` should log stderr and stdout outputs to `BuckEventBus`.
Summary:
If we have a shell step that outputs logs to stderr, the super console
basically freezes because the process executor prints directly to the
`Console`. This makes it so that a shell step NEVER asks the process
executor to print anything, but preserves the same behavior by logging
those messages to the buck event bus.

Test Plan:
Unit tests pass.