Add `buck project --target xcode`.
Summary:
This adds the project generator for xcode projects which handles
ios_library, xcode_native, and ios_test rules. It takes a build target,
and adds it and all its recursive dependencies into an xcode project. It
also generates a scheme that can be used to build the targets in the
right order.

Test Plan: Unit tests.