Separate ios_resource rule 'resources' into 'dirs' and 'files'.
Summary:
Yiding helped me realize that we need to treat directories separately
from files in ios_resource. We can handle directories correctly in
ios_resource rules by copying the directory and its contents to the
target location.

I noticed the constructor of `AppleResource` was incorrectly labeled
as `@VisibleForTesting`. It's used in a few places in the package,
so I removed the annotation.

Test Plan: Unit tests updated.