ProjectGenerator: Handle Buck build rules which do not map to Xcode targets.
Summary:
Yiding introduced a cache which maps Buck build rules to Xcode targets.
When I updated to include the cache, tests started failing because I
added support for a Buck build rule (genrule) which did not map to an
Xcode target.

After talking with Yiding, we decided to make the Xcode target optional,
and have clients check if it's present.