`source_under_test` for IosTestRule.
Summary:
This adds source_under_test attribute to `IosTestRule`. This field is used by
the xcode project generator to pull in tests related to targets. The project
generator now pulls in the named target, its dependencies, and the tests for
those, and their additional dependencies.

Test Plan: Added unit test.