Create a new buildable to filter drawable and string resources.
Summary:
This is the first part of making #aapt_resources and #uber_r_dot_java
buildables run parallel.

Right now, `UberRDotJava` does the following:
- Conditionally filter drawables (DPI specific) and strings (in asset builds).
- Generates R.java using directories filtered by the earlier step.
- Compiles and dexes the generated R.java.

`AaptPackageResources` depends on `UberRDotJava` because it packages the
resource directories filtered into the APK. Additionally, in the case of
strings-as-assets builds, it also creates those `.fbstr` string bundles,
for which it needs the R.txt/R.java files generated by `UberRDotJava`.

This diff pulls out the filtering step from `UberRDotJava` as a separate
buildable and adds that as dependency to the other two buildables. In a
subsequent diff, the plan is to pull out the logic that creates `.fbstr`
bundles into `AndroidBinaryRule` and package the assets using
`ApkBuilder`.