Add native implementation of exopackage installer agent
Summary:
It turns out that running dalvikvm on the Android command line isn't as
reliable as I had hoped.  One one device, it got an NPE during VM init.
On one emulator, it printed a warning about text relocations that
confused our output parsing (since ADB does not separate stdout from
stderr.)

The solution is to just do a native implementation, using the
executable-in-apk trick introduced in the last diff.  This native agent
is mostly identical to the Java version, and we can still switch back
and forth.

I use the public-domain miniz library to extract JAR signatures.  I have
disabled compression and a few other features to make it as small as
possible.

Test Plan: Installed exopackages on a few devices.