Easy: Fix typo in `AppleSource.ofSourcePath()`.
Summary:
I typo'd the enum value in `AppleSource.ofSourcePath()`. When I went to
use it in a subsequent diff, this bug smacked me in the face.

To ensure we cover this case in a test, I updated the constructor for
AppleSource to switch on the type enum and check the arguments
appropriately for the type.

Test Plan:
Ran TypeCoercerTest. It passes.  Note: `buck test --all` fails in master
with the same error, @natthu is working on it.