Clean up build file parse error reporting.
Summary:
Previously errors reported due to parse errors in build files were messy
and difficult to decipher. This change threads the `Console` through to
the `ProjectBuildFileParser` so it correctly disables `SuperConsole`
output when a parse error occurs, catches IllegalStateExceptions to
suppress stack trace printing and adds an integration test to check for
syntax error message output.