New utility class GroupedSources.
Summary:
I'm refactoring IosLibrary and friends to consolidate the 'srcs' and
'headers' field.  As part of this, I want to make the tree of
GroupedSource objects the source of truth for the list of SourcePath
objects belonging to an IosLibrary etc. rule.

This diff adds a utility class `GroupedSources` with a helper function
`GroupedSources`.sourcePaths() to walk one or more `GroupedSources`
trees and extract the `SourcePath` objects into an immutable collection.

We'll use this from `IosLibrary.getInputsToCompareToOutput()` and
friends.

Test Plan: Unit tests included.