Pass the buildId down to JUnitStep
Summary:
Pass buck's `buildId` down into `JUnitStep`, and from there into the JUnit processes themselves.

If both buck and the JUnit test include the `buildId` when they interact with an external resource, then the external resource can link the events together.

The practical example of this would be for buck to log the `buildId` into the Facebook Test Console when recording a test run, and for JUnit tests to log the `buildId` alongside any test users they create.  Once this is done, we can find all test users associated with a particular test run.

Test Plan: Dump the system property from a JUnit test, to make sure it's actually there.