Integration test for external javac
Summary:
Added an integration test that checks the full cycle of providing an
external javac in a custom project and checking the failing output.

Note: I had to run buck as a separate process because (according to
stackoverflow) I cannot change the current working directory in java. The
external javac is called with relative paths to the test project which are
meaningless in the toplevel buck directory.

Depends on D1212294

Test Plan: ##buck test --all##