Fix leaky thread pool abstractions.
Summary:
Previously the ThreadPools used to execute builds in parallel
were created by BuildContexts, but used by StepRunners and
AbstractCachingBuildRules and shut down by clients of those classes,
making the abstractions leaky and error prone. This change makes
DefaultStepRunner fully encapsulate the ThreadPool it uses and makes
it responsible for creating and shutting down the
pool. DefaultStepRunner and the Builds and BuildCommands that may
contain it are made Closable resources allowing clients of any of
those classes to use try-with-resources blocks to automatically close
the resources instead of having to know the details of how to reach in
to the object graph to manually close down thread pools and executors after
use.

Test Plan: buck test --all