Buck wrapper script: Don't read from stdin when not interactive
Summary:
Thanks to @vsi for reporting this. Previously, if you added a
file in the buck repo without checking it in, the wrapper script would
helpfully tell you about it and ask you if you wanted to clean the repo
by reading from stdin.

Unfortunately, since we call the wrapper script from another wrapper
script in fbobjc, this read never completed. We don't really care
about this case, so we'll just ignore it if stdin is not pointing
to a terminal.

Test Plan:
Created file in buck repo but didn't check in.
Ran 'buck -v', was prompted to clean repo. Repeated test, but
ran fbobjc's Tools/buck-project.py Libraries/FBProvider. Was
not prompted to clean repo and hang no longer reproduced.