Change `buck` to find the nearest `.buckconfig` file and use as project dir.
Summary:
The goal is to make it easier to run `buck` from any directory in a Buck project.
Instead of failing if there is no `.buckconfig` file in the working directory,
Buck now checks parent directories until it finds a `.buckconfig` file,
and uses that directory as the project root when running Buck.

Test Plan:
Try running `buck --version` from a subdirectory of a Buck project and ensure
that it runs successfully.
Try running `buck --version` from a parent directory of a Buck project and
ensure that the error message prints out.