Switch BuildTargetSourcePaths to be BuildRuleSourcePaths
Summary:
When using a SourcePath as an input to a RuleKey, it's beneficial to include not only the path to the
output file or build target as a string, but to hash the output file. First step to making this less
than equisitly painful to implement is to convert BuildTargetSourcePath to a BuildRuleSourcePath.
Common usage of this class tended to grab the target and resolve the rule, so there's less impact than
might be expected.

Test Plan:
buck test --all
Build the prod apps. Rebuild them. Make sure that the rule key isn't borked.