UberRDotJava should cache the classes.txt file instead of the total hash.
Summary:
Because SplitZipStep will need access to this map to compute the hash of its
inputs. I was only focussed on PreDexMerge until my last diff, which packs an
entire DexWithClasses and so doesn't need to know the hash of individual classes.

Sorry for the churn, but I want to keep the diffs small, because threading this
all the way from class path entries to the smart dexing step is going to be a
big diff itself.

Test Plan:
buck test --all
sandcastle